first version of backdoor
copy config.py.dist to config.py and change the values

the bot can be triggered by !backdoor PORT where port is a listening
port on the specified server. the backdoor server needs to have
socat listening on the port like

    socat file:`tty`,raw,echo=0 tcp-listen:4567

where 4567 is the port.

techniques taken from https://blog.ropnop.com/upgrading-simple-shells-to-fully-interactive-ttys/