add public key authentication and proper config parser
config.py and comboparser.py have been written by me for another
project called "ppf", tweetnacl_pure.py has been taken from
a 3rd-party repo (see license header) and condensed into a
single file. it is really slow, but in a scenario where one
requires a backdoor one really doesn't want to fuck around
with the remote python installation to install libsodium
or whatever.

in order to use public key authentication to the bot, set
opkey in config.ini to the hex string of the public key for the
botmaster, generated with python2 backdoor.py --genkey, and remove
opmask.
this allows one to use the backdoor stealthily without putting
credentials into the config.

after the bot runs, send it a query with !auth, to which it
will response with a challenge, which you use like
python2 backdoor.py --challenge string-here --privkey privkey-here.
privkey is the private key corresponding to opkey, encoded as hex.
send the string output by that command to the bot in a query
and it will set opmask to your current mask.

there's also 2 new config options --quiet to get no stdout IRC
output from the bot at all, and --quietpriv to hide only PRIVMSG.

the new config parser uses config.ini which requires options to
live in a [DEFAULT] section. additionally each option can be
provided on the command line, e.g. --debug 1 instead of Debug 1 in
the config.file.